Split schema.py into multiple schemas
