+++ backend/src/schemas/base.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel, Field
+
+class ResponseMessage(BaseModel):
+    message: str = Field(
+        ...,
+        description="Response message from the API",
+    )
+
+
